refactor(#234): remove internal 'StrictlyPositiveDoubleException'
